+++ lib/scale_design.dart
@@ -1,18 +1,23 @@
 library scale_design;
 
-export './src/size.dart'
+export 'src/size/size.dart'
     show Scale, scaleHeight, scaleWidth, scaleFont, perWidth, perHeight;
 
-export './src/widgets/text/t.dart' show T;
-export './src/widgets/text/t_span.dart' show TSpan;
-export './src/widgets/text/slide_text.dart' show SlideText;
+export 'src/size/widgets/text/t.dart' show T;
+export 'src/size/widgets/text/t_span.dart' show TSpan;
+export 'src/size/widgets/text/slide_text.dart' show SlideText;
 
-export './src/widgets/button/elevated_btn.dart' show ElevatedBtn;
-export './src/widgets/button/floating_action_btn.dart' show FloatingActionBtn;
-export './src/widgets/button/icon_btn.dart' show IconBtn;
-export './src/widgets/button/outlined_btn.dart' show OutlinedBtn;
-export './src/widgets/button/text_btn.dart' show TextBtn;
-export './src/widgets/button/circular_icon.dart' show CircularIcon;
-export 'src/components/scrollable_Icons_card.dart' show ScrollableIconsCard;
-export './src/components/toast.dart' show Toast;
-export './src/widgets/button/shiny_button.dart' show ShinyButton;
+export 'src/size/widgets/button/elevated_btn.dart' show ElevatedBtn;
+export 'src/size/widgets/button/floating_action_btn.dart'
+    show FloatingActionBtn;
+export 'src/size/widgets/button/icon_btn.dart' show IconBtn;
+export 'src/size/widgets/button/outlined_btn.dart' show OutlinedBtn;
+export 'src/size/widgets/button/text_btn.dart' show TextBtn;
+export 'src/size/widgets/button/circular_icon.dart' show CircularIcon;
+export 'src/size/components/scrollable_Icons_card.dart'
+    show ScrollableIconsCard;
+export 'src/size/components/toast.dart' show Toast;
+export 'src/size/widgets/button/shiny_button.dart' show ShinyButton;
+
+export 'src/responsive/responsive.dart';
+export 'src/responsive/responsive_nav_bar.dart';