+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 All notable changes will be documented in this file.
 
+## [0.0.9] - 20240301
+
+### Added
+
+- Added the `Responsive` component, which is a responsive layout tool.
+- Added the `ResposiveScaffold` component, which is a `Scaffold` for quickly building cross-platform navigation.
+
 ## [0.0.8] - 20231110
 
 ### Added
 
 - Add `ShinyButton` widget, which is a button with animated light and shadow effect.
-
 
 ## [0.0.4] - 20231104